Pagination

+++ owaincodes.Core/Models/PaginationDetails.cs
@@ -3,11 +3,12 @@
     public class PaginationDetails
     {
         public long CurrentPage { get; set; }
-        public long CurrentPageQty { get; set; }
+        public int CurrentPageQty { get; set; }
         public long PageSize { get; set; }
-        public long LastPage { get; set; }
+        public int LastPage { get; set; }
         public long TotalResults { get; set; }
 
         public string ItemType { get; set; }
     }
 }
+
